requirements/prod.txt
-r base.txt
raven==6.10.0
django-compressor==2.3
gunicorn==19.10.0
